Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
 
 int main() {
-    int a, b, c;
+    float a, b, c;
     
     // Taking input
     scanf("%d %d %d", &a, &b, &c);
 
-    int average = (a + b + c) / 3; // Integer division
+    float average = (a + b + c) / 3; // Integer division
 
     // Print result with newline
-    printf("Average: %d\n", average);
+    printf("Average: %.2f\n", average);
 
     return 0;
 }
